// src/lib/enums/products.enum.ts
export enum ProductSize {
    GB_16 = "16GB",
    NORMAL = "NORMAL",
    GB_32 = "32GB",
    GB_64 = "64GB",
    GB_128 = "128GB",
    GB_256 = "256GB",
    GB_512 = "512GB",
    TB_1 = "1TB",
    OTHER = "Other",
}

export enum TariffType {
    PREPAID = "PREPAID",
    POSTPAID = "POSTPAID",
    ROAMING = "ROAMING",
    CORPORATE = "CORPORATE",
    STUDENT = "STUDENT",
    FAMILY = "FAMILY",
}

export enum ProductBrand {
    SAMSUNG = "SAMSUNG",
    APPLE = "APPLE",
    LGU = "LGU",
    KT = "KT",
    SKT = "SKT",
    OTHER = "OTHER", // "Other" emas!
}

export enum ProductStatus {
    PAUSE = "PAUSE",
    PROCESS = "PROCESS",
    DELETE = "DELETE",
}
export enum ProductCollection {
    SMARTPHONE = "SMARTPHONE",
    TABLET = "TABLET",
    TARIFF = "TARIFF",
    AUDIO = "AUDIO",
    WEARABLE = "WEARABLE",
    CHARGER = "CHARGER",
    CASE = "CASE",
    SCREEN_PROTECTOR = "SCREEN_PROTECTOR",
    OTHER = "OTHER"
}